+++ tests/integration/distribution/keeper/grpc_query_test.go
@@ -49,11 +49,11 @@ func TestGRPCParams(t *testing.T) {
 					CommunityTax:        math.LegacyNewDecWithPrec(3, 1),
 					WithdrawAddrEnabled: true,
 					NakamotoBonus: types.NakamotoBonus{
-						Enabled:            true,
-						Step:               types.DefaultNakamotoBonusStep,
-						Period:             types.DefaultNakamotoBonusPeriod,
-						MinimumCoefficient: types.DefaultNakamotoBonusMinimumCoefficient,
-						MaximumCoefficient: types.DefaultNakamotoBonusMaximumCoefficient,
+						Enabled:               true,
+						Step:                  types.DefaultNakamotoBonusStep,
+						PeriodEpochIdentifier: types.DefaultNakamotoBonusPeriodEpochIdentifier,
+						MinimumCoefficient:    types.DefaultNakamotoBonusMinimumCoefficient,
+						MaximumCoefficient:    types.DefaultNakamotoBonusMaximumCoefficient,
 					},
 				}