+++ x/distribution/keeper/abci.go
@@ -14,11 +14,6 @@ func (k Keeper) BeginBlocker(ctx context.Context) error {
 	defer telemetry.ModuleMeasureSince(types.ModuleName, telemetry.Now(), telemetry.MetricKeyBeginBlocker)
 	c := sdk.UnwrapSDKContext(ctx)
 
-	// dynamically adjust the nakamoto bonus coefficient first.
-	if err := k.AdjustNakamotoBonusCoefficient(c); err != nil {
-		return err
-	}
-
 	// determine the total power signing the block
 	var previousTotalPower int64
 	for _, voteInfo := range c.VoteInfos() {