+++ x/distribution/keeper/allocation_test.go
@@ -76,7 +76,6 @@ func setupTestKeeper(t *testing.T, nakamotoBonusCoefficient math.LegacyDec, heig
 	} else {
 		require.NoError(t, err)
 	}
-	params.NakamotoBonus.Period = height
 	require.NoError(t, distrKeeper.Params.Set(ctx, params))
 
 	err = distrKeeper.SetNakamotoBonusCoefficient(ctx, nakamotoBonusCoefficient)